force float for floating point computation
Unless explicitly specified a floating point literal is a double:

float a = 0.2;
float b = 1.7 * a;

1.7 * a will be computed using double and not float, we should use 1.7f
here.

On 32bit architecture float are performed in hardware and double in
software, so using double is more expensive than double.

We should use cosf() instead of cos() (same for others) and 1.5f
instead of 1.5.

+++ src/graphics/camera.c
@@ -54,10 +54,10 @@ void camera_update_matrix(Camera *c, int width, int height)
 
 	angle = c->angle;
 	ratio = (float) width / height;
-	c->matrix[0] = cos(angle);
-	c->matrix[1] = sin(angle) * ratio;
-	c->matrix[2] = -sin(angle) / ratio;
-	c->matrix[3] = cos(angle);
+	c->matrix[0] = cosf(angle);
+	c->matrix[1] = sinf(angle) * ratio;
+	c->matrix[2] = -sinf(angle) / ratio;
+	c->matrix[3] = cosf(angle);
 }
 
 void camera_reset(Camera *c)
